Replace oorandom with rand crate
I have done this because oorandom was producing weird distortions and
‘patches’ that would definitely not appear in true randomness. At first
I thought it was oorandom’s implementation, but later I realised that it
was actually the fact that the RNG created for each pixel has the same
seed! Because I like oorandom’s minimalist approach, I tried using the
getrandom crate to get a random seed from the OS’s randomness source (as
suggested by the author of oorandom). Unfortunately this ended up being
around four times slower. At this point I relented to using the rand
crate, which ended up being around 8% slower (but in my opinion produces
images that are more than 8% better than those with the previous setup).